Add experimental grid class
This commit includes grid.py, which contains a start at refactoring the grid
type. Eventually the Grid will be a common basetype. The subclasses RegularGrid
and StructuredGrid encompass basic grid types useful for raster data. Future
work will involve bringing the RegularGrid fully up to par with AAIGrid and
adding basic operations to StructuredGrid (likely based on scipy).

AAIGrid remains independent for now, until this is more stable and RegularGrid
can be added as a super.