add note to update setuptools and modify travis script
It seems that since pip import setuptools automatically, setup.py is prevented
from updating the version of setuptools in place. One option is to lower the
required version of setuptools, but I'm not sure what the implications are.

Instead, I'm noting for now that setuptools are a requirement that must be
updated manually before using pip.