add experimental rewrite of karta.crs
I found the CRSRegistry to be annoying and pointless. Also, the
distinction between projected and local systems seems unimportant.

The new system discards the CRSRegistry, and the idiom will instead to
use a singly CRS instance across geometries. Custom objects are
implemented for cartesian planes, and a custom object will be
implemented later for spherical systems.