integrate quadtrees into Multipoint
- before using a quadtree, the `build_quadtree` method must be used (for
  now)

- currently, only `within_bbox` takes advantage of thie quadtree

- methods that modify the geometry destroy its quadtree (see below)

this doesn't have to be the case, but for point deletion the way
quadtree indices are handled looks expensive. Right now, the quadtree
contains an index internally for every point. Removing any point except
for the last would require the quadtree to be traversed for every index
larger than the removed point and the index decremented.