allow `Grid.sample` to take Point, Multipoint
also slightly changes the behavious of the `crs` keyword. If a vector
geometry object is provided, the crs is always taken from that. If
coordinate lists are provided (the old way), the crs is used, or if None
the crs is assumed the same as the grid crs (which should behave the
same as before).