override empty data arg in multipart constructor
geojson reader provides an empty dictionary when no "data-like" property
attributes are found, but Multipart takes a non-None dictionary as an
argument for creating a Table.

When provided a dictionary with length different than the vertices, an
inconsistent geometry is created, which will fail if indexed past the
length of the Table instance.

New behaviour is to replace an empty dictionary with an empty table of
the correct length, and raise ValueError when a non-empty dictionary
doesn't match the number of vertices

Furthermore, lists are no longer accepted as a value to the *data*
keyword argument. An AttributeError will be raised because lists lack a
`.items()` method.