short-term fix to grid indexing inconsistencies
Grid indexing with multiple bands needs some work. This changeset
improves consistency by adhering to the following rule:
- indexing returns a deep array unless the final dimension is singleton,
  except when the indexing key was two-dimensional, in which case the
  result is full-depth

This means that some operations which previously results in a
two-dimensional array, such as

    regular_grid[:,:]

now give a three-dimensional array.

This particular implementation is temporary, as bands and the
BandIndexer are soon to be overhauled